pyproject.toml
[project]
name = "replace-fonts"
version = "0.0.0"  # Formal value only; actual version managed in __version__
dependencies = [
    "python-pptx",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pytest",
    "ptpython",
    "types-pyyaml",
]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
show_error_codes = true

[tool.pytest]
addopts = ["-p", "no:cacheprovider", "-ra"]
testpaths = ["test"]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = [
    "A",      # flake8-builtins
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "E",      # pycodestyle errors
    "ERA",    # eradicate (commented-out code)
    "F",      # pyflakes
    "I",      # isort (import order)
    "N",      # pep8-naming
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "S",      # flake8-bandit
    "SIM",    # flake8-simplify
    "UP",     # pyupgrade (includes f-string)
    "W",      # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"test/**/*.py" = [
    "S101",  # Use of assert in tests is expected
    "S108",  # Hardcoded temporary file paths in test patterns
]

[tool.uv]
exclude-newer = "3 days"
